cdk/requirements.txt
-e .
aws_cdk.aws_s3
aws_cdk.aws_iam
aws_cdk.aws_kms
aws_cdk.aws_s3_deployment
aws_cdk.aws_wafv2
aws_cdk.aws_cloudfront
